move linked list into other files
